* rcs.c (RCS_delete_revs):  RCS_delete_revs uses an 	RCS_checkout call to get a new copy of a revision to be 	used internally after old revisions were deleted and it was 	performing keyword substitutions.  This munged all the 	the revisions of the file on the branch containing the 	deleted revisions and its sub-branches, as the original they 	were being patched from was incorrect.  Corrected this by 	passing in "-ko" as an option to RCS_checkout. 	* sanity.sh (keywordlog):  modified this test to verify the 	correct behavior of 'cvs admin -o'. 	[Fixed use of \$ in keywordlog test; added code in RCS_delete_revs 	to abort on binary file on Windows -kingdon]
CVS-ID: src/ChangeLog 1.1739
CVS-ID: src/rcs.c 1.216
CVS-ID: src/sanity.sh 1.531

+++ src/sanity.sh
@@ -15166,6 +15166,31 @@ xx"
 1\.5          (${username} *[0-9a-zA-Z-]*): change"
 	  cd ../..
 
+	  #
+	  # test the operation of 'admin -o' in conjunction with keywords
+	  # (especially Log - this used to munge the RCS file for all time)
+	  #
+
+	  dotest keywordlog-24 \
+"${testcvs} admin -oHEAD 1/first-dir/file1" \
+"RCS file: ${TESTDIR}/cvsroot/first-dir/file1,v
+deleting revision 1\.5
+done"
+
+	  dotest keywordlog-25 \
+"${testcvs} -q co -p first-dir/file1" \
+"initial
+xx "'\$'"Log: file1,v "'\$'"
+xx Revision 1\.4  [0-9/]* [0-9:]*  ${username}
+xx First log line
+xx Second log line
+xx"
+
+	  if test $keep = yes; then
+	    echo Keeping ${TESTDIR} and exiting due to --keep
+	    exit 0
+	  fi
+
 	  rm -r 1 2
 	  rm -rf ${CVSROOT_DIRNAME}/first-dir
 	  ;;
